backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
alembic
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
psycopg2-binary
pandas
python-dotenv
